Update z-wave.markdown
Folks that jump to this page from having installed the RPI image have no idea whatsoever that they have to do all of this from the virt env that HASS uses. Either use what I made or rephrase it so that it is clear at the beginning of these instructions that commands have to be ran from the virtual env.